+++ service/src/main/java/bio/terra/externalcreds/services/ProviderService.java
@@ -7,20 +7,10 @@ import bio.terra.externalcreds.auditLogging.AuditLogEventType;
 import bio.terra.externalcreds.auditLogging.AuditLogger;
 import bio.terra.externalcreds.config.ExternalCredsConfig;
 import bio.terra.externalcreds.config.ProviderProperties;
-import bio.terra.externalcreds.models.AuthorizationChangeEvent;
 import bio.terra.externalcreds.models.LinkedAccount;
 import bio.terra.externalcreds.models.LinkedAccountWithPassportAndVisas;
 import bio.terra.externalcreds.models.PassportVerificationDetails;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.api.core.ApiFutureCallback;
-import com.google.api.core.ApiFutures;
-import com.google.cloud.pubsub.v1.Publisher;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.util.concurrent.MoreExecutors;
-import com.google.protobuf.ByteString;
-import com.google.pubsub.v1.PubsubMessage;
-import java.io.IOException;
 import java.sql.Timestamp;
 import java.time.Duration;
 import java.time.Instant;
@@ -28,8 +18,6 @@ import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.TimeUnit;
-import javax.annotation.PreDestroy;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
@@ -53,8 +41,6 @@ public class ProviderService {
   private final PassportService passportService;
   private final JwtUtils jwtUtils;
   private final AuditLogger auditLogger;
-  private final ObjectMapper objectMapper;
-  private final Optional<Publisher> authorizationChangeEventPublisher;
 
   public ProviderService(
       ExternalCredsConfig externalCredsConfig,
@@ -63,9 +49,7 @@ public class ProviderService {
       LinkedAccountService linkedAccountService,
       PassportService passportService,
       JwtUtils jwtUtils,
-      AuditLogger auditLogger,
-      ObjectMapper objectMapper)
-      throws IOException {
+      AuditLogger auditLogger) {
     this.externalCredsConfig = externalCredsConfig;
     this.providerClientCache = providerClientCache;
     this.oAuth2Service = oAuth2Service;
@@ -73,20 +57,6 @@ public class ProviderService {
     this.passportService = passportService;
     this.jwtUtils = jwtUtils;
     this.auditLogger = auditLogger;
-    this.objectMapper = objectMapper;
-
-    // note that Publisher authenticates to Google using the env var GOOGLE_APPLICATION_CREDENTIALS
-    this.authorizationChangeEventPublisher =
-        externalCredsConfig
-            .getAuthorizationChangeEventTopicName()
-            .map(
-                topicName -> {
-                  try {
-                    return Publisher.newBuilder(topicName).build();
-                  } catch (IOException e) {
-                    throw new ExternalCredsException("exception building event publisher", e);
-                  }
-                });
   }
 
   public Set<String> getProviderList() {
@@ -373,46 +343,5 @@ public class ProviderService {
         linkedAccount.getUserId(),
         linkedAccount.getProviderId(),
         responseBody);
-  }
-
-  private void publishAuthorizationChangeEvent(AuthorizationChangeEvent event) {
-    authorizationChangeEventPublisher.ifPresent(
-        publisher -> {
-          try {
-            var message =
-                PubsubMessage.newBuilder()
-                    .setData(ByteString.copyFromUtf8(objectMapper.writeValueAsString(event)))
-                    .build();
-            var apiFuture = publisher.publish(message);
-            ApiFutures.addCallback(
-                apiFuture,
-                new ApiFutureCallback<>() {
-                  @Override
-                  public void onFailure(Throwable throwable) {
-                    log.error("failure publishing authorization change event", throwable);
-                  }
-
-                  @Override
-                  public void onSuccess(String messageId) {}
-                },
-                MoreExecutors.directExecutor());
-          } catch (JsonProcessingException e) {
-            throw new ExternalCredsException(
-                "json exception writing authorization change event:" + event, e);
-          }
-        });
-  }
-
-  @PreDestroy
-  void shutdownPublisher() {
-    authorizationChangeEventPublisher.ifPresent(
-        publisher -> {
-          try {
-            publisher.shutdown();
-            publisher.awaitTermination(1, TimeUnit.MINUTES);
-          } catch (InterruptedException e) {
-            throw new ExternalCredsException("publisher shutdown interrupted", e);
-          }
-        });
   }
 }
